Made UseFetch customHook to fetchData Repeatedly
